feat: accept Doorkeeper bearer tokens on /mcp, drop email auth
McpController now tries a Doorkeeper OAuth bearer token first, falling
back to the CF Access service-token branch — the interactive
email-match branch is deleted outright rather than kept alongside it.
That branch relied on Cloudflare Access's own policy having already
vetted who could get a JWT in the first place; once that policy is
loosened to admit anyone, trusting an unverified email claim in it
would reopen exactly the "unverified email" account-takeover risk the
existing threat assessment (docs/threat_assessments/MCP_OAUTH_AUTH.md)
had marked eliminated. The service-token branch is unaffected — it's
a possession-based secret, not something the Access policy gates.

CfAccessAuthentication is trimmed to a private library of CF JWT
mechanics with no before_action of its own; the new McpAuthentication
concern is the single dispatcher McpController includes, trying
Doorkeeper then CF service-token in order. This keeps future removal
of the CF fallback a one-method deletion.

Also restricts Doorkeeper's access_token_methods to the Authorization
header only (no query/body param fallback) — beyond the usual
token-leakage-via-logs concern, checking doorkeeper_token at all
forces Rails to eagerly parse request params, which was raising on
malformed JSON bodies before McpController's own parser got a chance
to render a proper JSON-RPC -32700 error for them.
